Use usize for truncate method in place of u64

+++ core/io/mod.rs
@@ -23,7 +23,7 @@ pub trait File: Send + Sync {
     ) -> Result<Arc<Completion>>;
     fn sync(&self, c: Arc<Completion>) -> Result<Arc<Completion>>;
     fn size(&self) -> Result<u64>;
-    fn truncate(&self, len: u64, c: Completion) -> Result<Arc<Completion>>;
+    fn truncate(&self, len: usize, c: Completion) -> Result<Arc<Completion>>;
 }
 
 #[derive(Debug, Copy, Clone, PartialEq)]
